requirements.txt
asgiref==3.2.3
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django>-3.0.3,<3.1.0
django-rest-swagger==2.2.0
djangorestframework>3.9.2.2,<3.11
djangorestframework_simplejwt

